Avoid using atof() for integers

+++ PJISP - Zadaci - T3 i T4/07 Cene struje u regionu/cena_struje_u_regionu.c
@@ -36,7 +36,7 @@ int main(int argc, char **argm)
         FILE *upis = otvoriDatoteku(argm[3], "w");
 
         int i = ucitajPodatke(datoteka);
-        int potrosenoKWh = atof(argm[1]);
+        int potrosenoKWh = atoi(argm[1]);
 
         ukupnaCena(i, potrosenoKWh);
         upisiIzvestaj(upis, i);
